pyproject.toml
[tool.poetry]
name = "logus"
version = "0.1.0"
description = "A library for crafting, analyzing, and perfecting AI prompts"
authors = ["Dipankar Sarkar <me@dipankar.name>"]
readme = "README.md"
packages = [
    { include = "logus" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = { version = "^0.115.0", optional = true }
uvicorn = { version = "^0.30.6", optional = true }
anthropic = "^0.34.2"
openai = "^1.47.1"
groq = "^0.11.0"
pydantic = { version = "^2.9.2", optional = true }
python-dotenv = "^1.0.1"
jinja2 = { version = "^3.1.5", optional = true }
click = "^8.1.7"

[tool.poetry.extras]
web = ["fastapi", "uvicorn", "pydantic", "jinja2"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pre-commit = "^3.5.0"
black = "^24.8.0"
flake8 = "^7.1.1"
mypy = "^1.11.2"

[tool.poetry.scripts]
logus = "logus.cli:cli"
logus-web = "logus.web_cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
strict = true
